Put cel package examples in package cel_test
This makes the example code rendered by pkg.go.dev appear as a small
main, improving UX.
